+++ src/main/java/com/paneedah/weaponlib/Weapon.java
@@ -2,7 +2,6 @@ package com.paneedah.weaponlib;
 
 import akka.japi.Pair;
 import com.paneedah.mwc.network.messages.BlockHitMessage;
-import com.paneedah.weaponlib.BulletHoleRenderer.BulletHole;
 import com.paneedah.weaponlib.animation.ScreenShakeAnimation;
 import com.paneedah.weaponlib.animation.ScreenShakingAnimationManager;
 import com.paneedah.weaponlib.animation.SpecialAttachments;
@@ -14,7 +13,6 @@ import com.paneedah.weaponlib.crafting.*;
 import com.paneedah.weaponlib.model.Shell;
 import com.paneedah.weaponlib.render.WeaponSpritesheetBuilder;
 import com.paneedah.weaponlib.render.shells.ShellParticleSimulator.Shell.Type;
-import io.redstudioragnarok.redcore.vectors.Vector3D;
 import io.redstudioragnarok.redcore.vectors.Vector3F;
 import net.minecraft.block.material.Material;
 import net.minecraft.block.state.IBlockState;
@@ -32,6 +30,7 @@ import net.minecraft.util.SoundEvent;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.RayTraceResult;
 import net.minecraft.util.math.Vec3d;
+import net.minecraft.util.text.TextComponentString;
 import net.minecraft.util.text.TextComponentTranslation;
 import net.minecraft.util.text.TextFormatting;
 import net.minecraft.world.World;
@@ -1287,9 +1286,8 @@ public class Weapon extends Item implements PlayerItemInstanceFactory<PlayerWeap
 //        if(world.isRemote)
 //            ClientEventHandler.BULLET_HOLE_RENDERER.addBulletHole(new BulletHole(new Vector3D(position.hitVec.x, position.hitVec.y, position.hitVec.z), position.sideHit, 0.05));
 
-        if(builder.blockImpactHandler != null) {
+        if (!world.isRemote && builder.blockImpactHandler != null)
             builder.blockImpactHandler.onImpact(world, player, entity, position);
-        }
     }
 
     @Override
@@ -1416,15 +1414,16 @@ public class Weapon extends Item implements PlayerItemInstanceFactory<PlayerWeap
         instance.setMaxShots(result);
         String message;
         if(result == 1) {
-            message = net.minecraft.util.text.translation.I18n.translateToLocalFormatted("gui.firearmMode.semi");
+            message = I18n.format("gui.firearmMode.semi");
         } else if(result == Integer.MAX_VALUE) {
-            message = net.minecraft.util.text.translation.I18n.translateToLocalFormatted("gui.firearmMode.auto");
+            message = I18n.format("gui.firearmMode.auto");
         } else {
-            message = net.minecraft.util.text.translation.I18n.translateToLocalFormatted("gui.firearmMode.burst");
+            message = I18n.format("gui.firearmMode.burst");
         }
         LOG.debug("Changed fire mode of {} to {}", instance, result);
 
-        modContext.getStatusMessageCenter().addMessage(net.minecraft.util.text.translation.I18n.translateToLocalFormatted("gui.firearmMode", message), 1000);
+        if (instance.getPlayer() instanceof EntityPlayer)
+            ((EntityPlayer) instance.getPlayer()).sendStatusMessage(new TextComponentString(I18n.format("gui.firearmMode", message)), true);
 
         instance.getPlayer().playSound(modContext.getChangeFireModeSound(), 1, 1);
     }
@@ -1491,7 +1490,9 @@ public class Weapon extends Item implements PlayerItemInstanceFactory<PlayerWeap
 
             float ratio = (minZoom - zoom) / (minZoom - maxZoom);
 
-            modContext.getStatusMessageCenter().addMessage(net.minecraft.util.text.translation.I18n.translateToLocalFormatted("gui.currentZoom", Math.round(ratio * 100)), 800);
+            if (instance.getPlayer() instanceof EntityPlayer)
+                ((EntityPlayer) instance.getPlayer()).sendStatusMessage(new TextComponentString(I18n.format("gui.currentZoom", Math.round(ratio * 100))), true);
+
             instance.getPlayer().playSound(modContext.getZoomSound(), 1, 1);
             LOG.debug("Changed optical zoom to {}", instance.getZoom());
         } else {
@@ -1513,7 +1514,10 @@ public class Weapon extends Item implements PlayerItemInstanceFactory<PlayerWeap
             instance.setZoom(zoom);
 
             float ratio = (minZoom - zoom) / (minZoom - maxZoom);
-            modContext.getStatusMessageCenter().addMessage(net.minecraft.util.text.translation.I18n.translateToLocalFormatted("gui.currentZoom", Math.round(ratio * 100)), 800);
+
+            if (instance.getPlayer() instanceof EntityPlayer)
+                ((EntityPlayer) instance.getPlayer()).sendStatusMessage(new TextComponentString(I18n.format("gui.currentZoom", Math.round(ratio * 100))), true);
+
             instance.getPlayer().playSound(modContext.getZoomSound(), 1, 1);
             LOG.debug("Changed optical zoom to {}", zoom);
         } else {